20050610
git-svn-id: svn://sourceware.org/svn/prelink@125 94c539fb-cf18-0410-b60f-edeeb537fa16

+++ testsuite/reloc8.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 . `dirname $0`/functions.sh
+# Disable this test under SELinux
+test -x /usr/sbin/getenforce -a "`/usr/sbin/getenforce`" = Enforcing && exit 77
 rm -f reloc8 reloc8lib*.so reloc8.log
 rm -f prelink.cache
 NOCOPYRELOC=-Wl,-z,nocopyreloc
